requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.4
python-dotenv==1.0.0
aria2p==0.11.3
requests==2.31.0
tqdm==4.66.1
psutil==5.9.6
aiofiles==23.2.1
